enhancements/requirements.txt
anthropic>=0.52.0
boto3>=1.34.0
google-api-python-client>=2.160.0
google-auth>=2.38.0
gspread>=6.1.2
python-dotenv>=1.0.1
reportlab>=4.2.0
requests>=2.32.0
twilio>=9.3.0
